% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtetrad-package.R
\name{print.tetrad_result}
\alias{print.tetrad_result}
\title{Print a tetrad_result object}
\usage{
\method{print}{tetrad_result}(x, ...)
}
\arguments{
\item{x}{A \code{tetrad_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print a tetrad_result object
}
